python/arduino-health/requirements.txt
requests
pyserial==3.5